Add Basics of Data Manipulation in Time Series notebook
This commit introduces a new Jupyter notebook, titled "S1 Basics of Data Manipulation in Time Series". This notebook contains various operations to manipulate data in a dataframe, including importing data from an excel file, viewing a portion of the dataframe using the head and tail functions, and generating basic statistical summaries of the dataset's features.